\name{Rb}
\alias{Rb}
\docType{data}
\title{
Brazilian savannah background reflectance spectrum.}
\description{
Brazilian savannah background reflectance spectrum calculated by the average reflectance of leaf, leaf litter, tree bark and twigs.}
\usage{data("Rb")}
\format{
  A data frame with 401 observations on the following 2 variables.
  \describe{
    \item{\code{X300.700}}{a numeric vector}
    \item{\code{cerrado}}{a numeric vector}
  }
}
\source{
Gawryszewski, F. M., and P. C. Motta. 2012. Colouration of the orb-web spider Gasteracantha cancriformis does not increase its foraging success. Ethol Ecol Evol 24:23-38.}
\keyword{datasets}
